+++ api/src/main/kotlin/api/routing/stops/StopsClient.kt
@@ -1,15 +1,15 @@
 package api.routing.stops
 
-import api.db.*
-import api.db.models.LineDestination
-import api.db.models.toDeviceToken
-import api.extensions.bindJson
 import api.extensions.getWrapped
 import api.routing.Response.*
 import api.utils.Pipeline
 import api.utils.StopTimesF
 import arrow.core.raise.either
-import crtm.utils.createStopCode
+import common.extensions.bindJson
+import common.models.LineDestination
+import common.models.toDeviceToken
+import common.queries.*
+import common.utils.createStopCode
 import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
@@ -17,7 +17,7 @@ import io.ktor.server.plugins.cachingheaders.*
 import io.ktor.server.request.*
 import kotlinx.coroutines.flow.toList
 import simpleJson.*
-import api.db.getAllStops as getAllStopsFromDb
+import common.queries.getAllStops as getAllStopsFromDb
 
 
 fun Pipeline.getAllStops(): ResponseFlowJson {