+++ api/src/main/kotlin/api/routing/stops/train/TrainStopsRouting.kt
@@ -4,7 +4,7 @@ import api.routing.handle
 import api.routing.stops.alertsConfigF
 import api.routing.stops.getStopTimesResponse
 import api.routing.stops.subConfigF
-import api.routing.stops.trainRouted.trainCodMode
+import common.utils.trainCodMode
 import io.ktor.server.routing.*
 
 fun Route.trainStopsRouting() = route("/train") {