+++ api/src/main/kotlin/api/utils/TypeAliases.kt
@@ -1,8 +1,8 @@
 package api.utils
 
-import api.exceptions.BusTrackerException
 import api.routing.stops.StopTimes
 import arrow.core.Either
+import common.exceptions.BusTrackerException
 import io.ktor.server.application.*
 import io.ktor.util.pipeline.*