+++ loader/src/main/kotlin/loader/Main.kt
@@ -1,0 +1,9 @@
+package loader
+
+import common.DB
+import common.extensions.getOrThrow
+
+suspend fun main() {
+    DB.setupMongo(EnvVariables.mongoConnectionString.getOrThrow())
+    loadDataIntoDb()
+}